BotChooseLTGItem: float FPU temps remove 8-byte align prologue (-79 byte_diffs)
BotChangeViewAngles: 8-byte locals trigger MSVC6's and esp,0xfffffff8
prologue. With them gone, the function emits the bare sub esp,0x48
prologue matching the reference at 0x1002feb0.

FuzzyWeightUndecided returns double but the value lives entirely in
st7 throughout — no stack spill — so converting the declared type
only changes how MSVC reserves stack, not the precision used.

audit: MATCH=480 ICM=37 byte_diffs=165608 (was 165687, -79)
aarch64 build OK.

Source-commit: e1e5b7d83f7c0b6854598ece033ca44230144c23

+++ botlib/botlib.c
@@ -28592,8 +28592,8 @@ int __cdecl BotChooseLTGItem(int *a1, vec3_t a2, char *a3, int a4)
   int v9; // ebx
   iteminfo_t *v10; // edi (was _DWORD *) - 64-bit fix
   int v11; // eax
-  double v12; // st7
-  double v13; // st7
+  float v12; // st7 (was double)
+  float v13; // st7 (was double)
   levelitem_t *v14; // [esp+10h] [ebp-48h]
   int v15; // [esp+14h] [ebp-44h]
   itemconfig_t *v16; // [esp+18h] [ebp-40h]  - 64-bit fix (was int)
